requirements.lock.txt
aiohttp==3.8.1
aiosignal==1.2.0
ape-alchemy==0.6.0
ape-ens==0.6.0
ape-etherscan==0.6.0
ape-hardhat==0.6.0
ape-solidity==0.6.0
ape-tokens==0.6.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==1.0.3
bitarray==2.7.3
black==23.1.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
crytic-compile==0.2.3
cytoolz==0.11.2
dataclassy==0.11.1
decorator==5.1.1
Deprecated==1.2.13
devtools==0.10.0
eip712==0.2.0
eth-abi==3.0.1
eth-account==0.7.0
eth-ape==0.6.4
eth-bloom==1.0.4
eth-hash==0.3.2
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.7.0b1
eth-typing==3.1.0
eth-utils==2.0.0
ethpm-types==0.4.1
evm-trace==0.1.0a17
executing==1.2.0
flake8==5.0.4
frozenlist==1.3.0
greenlet==1.1.3
hexbytes==0.3.0
idna==3.3
ijson==3.1.4
importlib-metadata==4.11.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==8.11.0
jedi==0.18.1
jsonschema==4.13.0
lru-dict==1.1.7
matplotlib-inline==0.1.3
mccabe==0.7.0
morphys==1.0
msgspec==0.12.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parsimonious==0.8.1
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.0
pluggy==1.0.0
prettytable==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cid==0.3.0
py-ecc==6.0.0
py-evm==0.6.0a1
py-geth==3.11.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==1.1.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.10.5
pyethash==0.1.27
pyflakes==2.5.0
pygit2==1.9.1
PyGithub==1.55
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pysha3==1.0.2
pytest==6.2.5
python-baseconv==1.2.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rich==12.6.0
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
slither-analyzer==0.8.3
solc-select==0.2.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
stack-data==0.2.0
tokenlists==0.1.2
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tqdm==4.63.1
traitlets==5.3.0
trie==2.1.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.9
varint==1.0.2
watchdog==2.3.1
wcwidth==0.2.5
web3==6.0.0b7
websockets==10.3
wrapt==1.14.0
yarl==1.7.2
zipp==3.7.0
